Use build numbers in release changelog
This patch completely changes the algorithm for selecting the tags
used in the release changelog.

Before, it picked up all 7.x/6.x tags and the most recent tag.

This broke down during the review period. 7.3 was submitted for
review, a 7.4 beta shipped, but a 7.3 tag is only created AFTER
review. This made the top diff be between 7.2 and the 7.4 beta.

This patches changes the algorithm to select the most recent build
out of every 7.x/6.x cycle. It will totally break when the app store
build is NOT the most recent build, like if someone uploads a 7.3
build after 7.3 was released. I couldn't find a recent example of
that, but it's pretty flaky.